+++ src/app/result/result.component.ts
@@ -1,7 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-
-const MAX_SIZE = 1e5;
-const ERROR_MAX_OUTPUT_SIZE = 'Error: Output exceeded maximum size allowed';
+import { MAX_DATA_SIZE, ERROR_MAX_DATA_SIZE } from '../app.config';
 
 @Component({
   selector: 'app-result',
@@ -11,21 +9,29 @@ const ERROR_MAX_OUTPUT_SIZE = 'Error: Output exceeded maximum size allowed';
 export class ResultComponent implements OnInit {
   result: string;
   visibleResult: string;
+  limitReached = false;
 
   constructor() { }
 
   clean() {
     this.result = '';
     this.visibleResult = '';
+    this.limitReached = false;
   }
 
   addLine(line) {
+    // limit storing more data to avoid memory leaks
+    if (this.limitReached) {
+      return false;
+    }
+
     this.result += line + '\n';
 
-    if (this.result.length < MAX_SIZE) {
+    if (this.result.length < MAX_DATA_SIZE) {
       this.visibleResult = this.result;
     } else {
-      this.visibleResult += `${this.result.slice(0, MAX_SIZE)}\n\n${ERROR_MAX_OUTPUT_SIZE}`;
+      this.limitReached = true;
+      this.visibleResult += `${this.result.slice(0, MAX_DATA_SIZE)}\n\n${ERROR_MAX_DATA_SIZE}`;
       return false;
     }
   }