+++ changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Avoid memory leaks in case of much output data
+### Changed
+- Better Output stringify
 
 ## [1.0.1] - 2020-01-12
 ### Fixed